Create a separate hex::Error type
