+++ src/sha512.rs
@@ -114,8 +114,8 @@ impl hash::Hash for Hash {
 }
 
 impl str::FromStr for Hash {
-    type Err = Error;
-    fn from_str(s: &str) -> Result<Self, Error> {
+    type Err = ::hex::Error;
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
         ::hex::FromHex::from_hex(s)
     }
 }
